reorganize testing data functions
